django-oidc-bernie/bernie_project/requirements.txt
django>=1.8
https://github.com/juanifioren/django-oidc-provider/archive/v0.2.x.zip
mock>=1.0.1
# There is no OpenID lib that support both Python 2 and 3 from a single-source base.
# So we have to choose between:
# python-openid>=2.2.5
# python3-openid>=3.0.1
python-openid>=2.2.5
requests>=2.0.0
requests-oauthlib>=0.3.0
django-allauth>=0.13.0
django-bootstrap-form
